Working on Belfast feedback

+++ linear_algebra/code/include/linear_algebra/addition_traits.hpp
@@ -352,7 +352,7 @@ struct matrix_addition_engine_traits<OT, dr_vector_engine<T1, A1>, dr_vector_eng
 //--------------------------------------
 //- dr_vector_engine + fs_vector_engine.
 //
-template<class OT, class T1, class A1, class T2, int32_t N2>
+template<class OT, class T1, class A1, class T2, size_t N2>
 struct matrix_addition_engine_traits<OT, dr_vector_engine<T1, A1>, fs_vector_engine<T2, N2>>
 {
     using element_type = matrix_addition_element_t<OT, T1, T2>;
@@ -363,7 +363,7 @@ struct matrix_addition_engine_traits<OT, dr_vector_engine<T1, A1>, fs_vector_eng
 //--------------------------------------
 //- fs_vector_engine + dr_vector_engine.
 //
-template<class OT, class T1, int32_t N1, class T2, class A2>
+template<class OT, class T1, size_t N1, class T2, class A2>
 struct matrix_addition_engine_traits<OT, fs_vector_engine<T1, N1>, dr_vector_engine<T2, A2>>
 {
     using element_type = matrix_addition_element_t<OT, T1, T2>;
@@ -374,7 +374,7 @@ struct matrix_addition_engine_traits<OT, fs_vector_engine<T1, N1>, dr_vector_eng
 //--------------------------------------
 //- fs_vector_engine + fs_vector_engine.
 //
-template<class OT, class T1, int32_t N1, class T2, int32_t N2>
+template<class OT, class T1, size_t N1, class T2, size_t N2>
 struct matrix_addition_engine_traits<OT, fs_vector_engine<T1, N1>, fs_vector_engine<T2, N2>>
 {
     static_assert(N1 == N2);
@@ -431,7 +431,7 @@ struct matrix_addition_engine_traits<OT,
 //--------------------------------------
 //- dr_matrix_engine + fs_matrix_engine.
 //
-template<class OT, class T1, class A1, class T2, int32_t R2, int32_t C2>
+template<class OT, class T1, class A1, class T2, size_t R2, size_t C2>
 struct matrix_addition_engine_traits<OT,
                                      dr_matrix_engine<T1, A1>,
                                      fs_matrix_engine<T2, R2, C2>>
@@ -441,7 +441,7 @@ struct matrix_addition_engine_traits<OT,
     using engine_type  = dr_matrix_engine<element_type, alloc_type>;
 };
 
-template<class OT, class T1, class A1, class T2, int32_t R2, int32_t C2>
+template<class OT, class T1, class A1, class T2, size_t R2, size_t C2>
 struct matrix_addition_engine_traits<OT,
                                      dr_matrix_engine<T1, A1>,
                                      matrix_transpose_view<fs_matrix_engine<T2, R2, C2>>>
@@ -451,7 +451,7 @@ struct matrix_addition_engine_traits<OT,
     using engine_type  = dr_matrix_engine<element_type, alloc_type>;
 };
 
-template<class OT, class T1, class A1, class T2, int32_t R2, int32_t C2>
+template<class OT, class T1, class A1, class T2, size_t R2, size_t C2>
 struct matrix_addition_engine_traits<OT,
                                      matrix_transpose_view<dr_matrix_engine<T1, A1>>,
                                      fs_matrix_engine<T2, R2, C2>>
@@ -461,7 +461,7 @@ struct matrix_addition_engine_traits<OT,
     using engine_type  = dr_matrix_engine<element_type, alloc_type>;
 };
 
-template<class OT, class T1, class A1, class T2, int32_t R2, int32_t C2>
+template<class OT, class T1, class A1, class T2, size_t R2, size_t C2>
 struct matrix_addition_engine_traits<OT,
                                      matrix_transpose_view<dr_matrix_engine<T1, A1>>,
                                      matrix_transpose_view<fs_matrix_engine<T2, R2, C2>>>
@@ -474,7 +474,7 @@ struct matrix_addition_engine_traits<OT,
 //--------------------------------------
 //- fs_matrix_engine + dr_matrix_engine.
 //
-template<class OT, class T1, int32_t R1, int32_t C1, class T2, class A2>
+template<class OT, class T1, size_t R1, size_t C1, class T2, class A2>
 struct matrix_addition_engine_traits<OT,
                                      fs_matrix_engine<T1, R1, C1>,
                                      dr_matrix_engine<T2, A2>>
@@ -484,7 +484,7 @@ struct matrix_addition_engine_traits<OT,
     using engine_type  = dr_matrix_engine<element_type, alloc_type>;
 };
 
-template<class OT, class T1, int32_t R1, int32_t C1, class T2, class A2>
+template<class OT, class T1, size_t R1, size_t C1, class T2, class A2>
 struct matrix_addition_engine_traits<OT,
                                      fs_matrix_engine<T1, R1, C1>,
                                      matrix_transpose_view<dr_matrix_engine<T2, A2>>>
@@ -494,7 +494,7 @@ struct matrix_addition_engine_traits<OT,
     using engine_type  = dr_matrix_engine<element_type, alloc_type>;
 };
 
-template<class OT, class T1, int32_t R1, int32_t C1, class T2, class A2>
+template<class OT, class T1, size_t R1, size_t C1, class T2, class A2>
 struct matrix_addition_engine_traits<OT,
                                      matrix_transpose_view<fs_matrix_engine<T1, R1, C1>>,
                                      dr_matrix_engine<T2, A2>>
@@ -504,7 +504,7 @@ struct matrix_addition_engine_traits<OT,
     using engine_type  = dr_matrix_engine<element_type, alloc_type>;
 };
 
-template<class OT, class T1, int32_t R1, int32_t C1, class T2, class A2>
+template<class OT, class T1, size_t R1, size_t C1, class T2, class A2>
 struct matrix_addition_engine_traits<OT,
                                      matrix_transpose_view<fs_matrix_engine<T1, R1, C1>>,
                                      matrix_transpose_view<dr_matrix_engine<T2, A2>>>
@@ -517,7 +517,7 @@ struct matrix_addition_engine_traits<OT,
 //--------------------------------------
 //- fs_matrix_engine + fs_matrix_engine.
 //
-template<class OT, class T1, int32_t R1, int32_t C1, class T2, int32_t R2, int32_t C2>
+template<class OT, class T1, size_t R1, size_t C1, class T2, size_t R2, size_t C2>
 struct matrix_addition_engine_traits<OT,
                                      fs_matrix_engine<T1, R1, C1>,
                                      fs_matrix_engine<T2, R2, C2>>
@@ -528,7 +528,7 @@ struct matrix_addition_engine_traits<OT,
     using engine_type  = fs_matrix_engine<element_type, R1, C1>;
 };
 
-template<class OT, class T1, int32_t R1, int32_t C1, class T2, int32_t R2, int32_t C2>
+template<class OT, class T1, size_t R1, size_t C1, class T2, size_t R2, size_t C2>
 struct matrix_addition_engine_traits<OT,
                                      matrix_transpose_view<fs_matrix_engine<T1, R1, C1>>,
                                      fs_matrix_engine<T2, R2, C2>>
@@ -539,7 +539,7 @@ struct matrix_addition_engine_traits<OT,
     using engine_type  = fs_matrix_engine<element_type, R2, C2>;
 };
 
-template<class OT, class T1, int32_t R1, int32_t C1, class T2, int32_t R2, int32_t C2>
+template<class OT, class T1, size_t R1, size_t C1, class T2, size_t R2, size_t C2>
 struct matrix_addition_engine_traits<OT,
                                      fs_matrix_engine<T1, R1, C1>,
                                      matrix_transpose_view<fs_matrix_engine<T2, R2, C2>>>
@@ -550,7 +550,7 @@ struct matrix_addition_engine_traits<OT,
     using engine_type  = fs_matrix_engine<element_type, R1, C1>;
 };
 
-template<class OT, class T1, int32_t R1, int32_t C1, class T2, int32_t R2, int32_t C2>
+template<class OT, class T1, size_t R1, size_t C1, class T2, size_t R2, size_t C2>
 struct matrix_addition_engine_traits<OT,
                                      matrix_transpose_view<fs_matrix_engine<T1, R1, C1>>,
                                      matrix_transpose_view<fs_matrix_engine<T2, R2, C2>>>
