+++ linear_algebra/code/include/linear_algebra/subtraction_traits.hpp
@@ -351,7 +351,7 @@ struct matrix_subtraction_engine_traits<OT, dr_vector_engine<T1, A1>, dr_vector_
 //--------------------------------------
 //- dr_vector_engine - fs_vector_engine.
 //
-template<class OT, class T1, class A1, class T2, int32_t N2>
+template<class OT, class T1, class A1, class T2, size_t N2>
 struct matrix_subtraction_engine_traits<OT, dr_vector_engine<T1, A1>, fs_vector_engine<T2, N2>>
 {
     using element_type = matrix_subtraction_element_t<OT, T1, T2>;
@@ -362,7 +362,7 @@ struct matrix_subtraction_engine_traits<OT, dr_vector_engine<T1, A1>, fs_vector_
 //--------------------------------------
 //- fs_vector_engine - dr_vector_engine.
 //
-template<class OT, class T1, int32_t N1, class T2, class A2>
+template<class OT, class T1, size_t N1, class T2, class A2>
 struct matrix_subtraction_engine_traits<OT, fs_vector_engine<T1, N1>, dr_vector_engine<T2, A2>>
 {
     using element_type = matrix_subtraction_element_t<OT, T1, T2>;
@@ -373,7 +373,7 @@ struct matrix_subtraction_engine_traits<OT, fs_vector_engine<T1, N1>, dr_vector_
 //--------------------------------------
 //- fs_vector_engine - fs_vector_engine.
 //
-template<class OT, class T1, int32_t N1, class T2, int32_t N2>
+template<class OT, class T1, size_t N1, class T2, size_t N2>
 struct matrix_subtraction_engine_traits<OT, fs_vector_engine<T1, N1>, fs_vector_engine<T2, N2>>
 {
     static_assert(N1 == N2);
@@ -430,7 +430,7 @@ struct matrix_subtraction_engine_traits<OT,
 //--------------------------------------
 //- dr_matrix_engine - fs_matrix_engine.
 //
-template<class OT, class T1, class A1, class T2, int32_t R2, int32_t C2>
+template<class OT, class T1, class A1, class T2, size_t R2, size_t C2>
 struct matrix_subtraction_engine_traits<OT,
                                         dr_matrix_engine<T1, A1>,
                                         fs_matrix_engine<T2, R2, C2>>
@@ -440,7 +440,7 @@ struct matrix_subtraction_engine_traits<OT,
     using engine_type  = dr_matrix_engine<element_type, alloc_type>;
 };
 
-template<class OT, class T1, class A1, class T2, int32_t R2, int32_t C2>
+template<class OT, class T1, class A1, class T2, size_t R2, size_t C2>
 struct matrix_subtraction_engine_traits<OT,
                                         dr_matrix_engine<T1, A1>,
                                         matrix_transpose_view<fs_matrix_engine<T2, R2, C2>>>
@@ -450,7 +450,7 @@ struct matrix_subtraction_engine_traits<OT,
     using engine_type  = dr_matrix_engine<element_type, alloc_type>;
 };
 
-template<class OT, class T1, class A1, class T2, int32_t R2, int32_t C2>
+template<class OT, class T1, class A1, class T2, size_t R2, size_t C2>
 struct matrix_subtraction_engine_traits<OT,
                                         matrix_transpose_view<dr_matrix_engine<T1, A1>>,
                                         fs_matrix_engine<T2, R2, C2>>
@@ -460,7 +460,7 @@ struct matrix_subtraction_engine_traits<OT,
     using engine_type  = dr_matrix_engine<element_type, alloc_type>;
 };
 
-template<class OT, class T1, class A1, class T2, int32_t R2, int32_t C2>
+template<class OT, class T1, class A1, class T2, size_t R2, size_t C2>
 struct matrix_subtraction_engine_traits<OT,
                                         matrix_transpose_view<dr_matrix_engine<T1, A1>>,
                                         matrix_transpose_view<fs_matrix_engine<T2, R2, C2>>>
@@ -473,7 +473,7 @@ struct matrix_subtraction_engine_traits<OT,
 //--------------------------------------
 //- fs_matrix_engine - dr_matrix_engine.
 //
-template<class OT, class T1, int32_t R1, int32_t C1, class T2, class A2>
+template<class OT, class T1, size_t R1, size_t C1, class T2, class A2>
 struct matrix_subtraction_engine_traits<OT,
                                         fs_matrix_engine<T1, R1, C1>,
                                         dr_matrix_engine<T2, A2>>
@@ -483,7 +483,7 @@ struct matrix_subtraction_engine_traits<OT,
     using engine_type  = dr_matrix_engine<element_type, alloc_type>;
 };
 
-template<class OT, class T1, int32_t R1, int32_t C1, class T2, class A2>
+template<class OT, class T1, size_t R1, size_t C1, class T2, class A2>
 struct matrix_subtraction_engine_traits<OT,
                                         fs_matrix_engine<T1, R1, C1>,
                                         matrix_transpose_view<dr_matrix_engine<T2, A2>>>
@@ -493,7 +493,7 @@ struct matrix_subtraction_engine_traits<OT,
     using engine_type  = dr_matrix_engine<element_type, alloc_type>;
 };
 
-template<class OT, class T1, int32_t R1, int32_t C1, class T2, class A2>
+template<class OT, class T1, size_t R1, size_t C1, class T2, class A2>
 struct matrix_subtraction_engine_traits<OT,
                                         matrix_transpose_view<fs_matrix_engine<T1, R1, C1>>,
                                         dr_matrix_engine<T2, A2>>
@@ -503,7 +503,7 @@ struct matrix_subtraction_engine_traits<OT,
     using engine_type  = dr_matrix_engine<element_type, alloc_type>;
 };
 
-template<class OT, class T1, int32_t R1, int32_t C1, class T2, class A2>
+template<class OT, class T1, size_t R1, size_t C1, class T2, class A2>
 struct matrix_subtraction_engine_traits<OT,
                                         matrix_transpose_view<fs_matrix_engine<T1, R1, C1>>,
                                         matrix_transpose_view<dr_matrix_engine<T2, A2>>>
@@ -516,7 +516,7 @@ struct matrix_subtraction_engine_traits<OT,
 //--------------------------------------
 //- fs_matrix_engine - fs_matrix_engine.
 //
-template<class OT, class T1, int32_t R1, int32_t C1, class T2, int32_t R2, int32_t C2>
+template<class OT, class T1, size_t R1, size_t C1, class T2, size_t R2, size_t C2>
 struct matrix_subtraction_engine_traits<OT,
                                         fs_matrix_engine<T1, R1, C1>,
                                         fs_matrix_engine<T2, R2, C2>>
@@ -527,7 +527,7 @@ struct matrix_subtraction_engine_traits<OT,
     using engine_type  = fs_matrix_engine<element_type, R1, C1>;
 };
 
-template<class OT, class T1, int32_t R1, int32_t C1, class T2, int32_t R2, int32_t C2>
+template<class OT, class T1, size_t R1, size_t C1, class T2, size_t R2, size_t C2>
 struct matrix_subtraction_engine_traits<OT,
                                         matrix_transpose_view<fs_matrix_engine<T1, R1, C1>>,
                                         fs_matrix_engine<T2, R2, C2>>
@@ -538,7 +538,7 @@ struct matrix_subtraction_engine_traits<OT,
     using engine_type  = fs_matrix_engine<element_type, R2, C2>;
 };
 
-template<class OT, class T1, int32_t R1, int32_t C1, class T2, int32_t R2, int32_t C2>
+template<class OT, class T1, size_t R1, size_t C1, class T2, size_t R2, size_t C2>
 struct matrix_subtraction_engine_traits<OT,
                                         fs_matrix_engine<T1, R1, C1>,
                                         matrix_transpose_view<fs_matrix_engine<T2, R2, C2>>>
@@ -549,7 +549,7 @@ struct matrix_subtraction_engine_traits<OT,
     using engine_type  = fs_matrix_engine<element_type, R1, C1>;
 };
 
-template<class OT, class T1, int32_t R1, int32_t C1, class T2, int32_t R2, int32_t C2>
+template<class OT, class T1, size_t R1, size_t C1, class T2, size_t R2, size_t C2>
 struct matrix_subtraction_engine_traits<OT,
                                         matrix_transpose_view<fs_matrix_engine<T1, R1, C1>>,
                                         matrix_transpose_view<fs_matrix_engine<T2, R2, C2>>>
